Move hue related files to a separate package
