+++ src/main/kotlin/com/larastudios/chambrier/adapter/hue/CommonTypes.kt
@@ -1,4 +1,4 @@
-package com.larastudios.chambrier.adapter
+package com.larastudios.chambrier.adapter.hue
 
 data class HueResponse<T>(
     val data: List<T>,
