+++ src/main/kotlin/com/larastudios/chambrier/adapter/hue/HueClient.kt
@@ -1,4 +1,4 @@
-package com.larastudios.chambrier.adapter
+package com.larastudios.chambrier.adapter.hue
 
 import org.springframework.stereotype.Service
 import org.springframework.web.reactive.function.client.WebClient
