+++ src/main/kotlin/com/larastudios/chambrier/adapter/hue/LightsResponse.kt
@@ -1,4 +1,4 @@
-package com.larastudios.chambrier.adapter
+package com.larastudios.chambrier.adapter.hue
 
 data class LightGet(
     val id: String,
